Add SUPEE-9881 patch to the repo
SUPEE-9881 fixes an issue with removing a coupon code from the
cart page. After entering a coupon code and removing it, the coupon
is re-added in the checkout.

The patch has been added for EE 1.14.3.2, but can also be applied for
other versions. The issue is introduced in 1.9.3.0 and 1.14.3.0 in
both the Community version and the Enterprise versions.